+++ lib/global/utils/findSeparatorGroup.ts
@@ -1,14 +1,14 @@
-import { assert } from "../../utils/assert";
 import { read } from "../mutableState";
 
 export function findSeparatorGroup(separatorElement: HTMLElement) {
-  const groupElement = separatorElement.parentElement;
-  assert(groupElement, "Parent group element not found");
-
   const { mountedGroups } = read();
 
   for (const [group] of mountedGroups) {
-    if (group.element === groupElement) {
+    if (
+      group.separators.some(
+        (separator) => separator.element === separatorElement
+      )
+    ) {
       return group;
     }
   }