+++ src/components/Link.tsx
@@ -3,10 +3,12 @@ import { Link as ExternalLink, type DefaultPath } from "react-lib-tools";
 import type { Path } from "../routes";
 
 export function Link({
+  anchor,
   to,
   ...rest
 }: HTMLAttributes<HTMLSpanElement> & {
+  anchor?: string;
   to: Path | DefaultPath;
 }) {
-  return <ExternalLink to={to} {...rest} />;
+  return <ExternalLink to={anchor ? `${to}#${anchor}` : to} {...rest} />;
 }